requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.4
django-bootstrap4==23.2
django-countries==7.5.1
django-filter==23.2
Pillow==10.0.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
